Create function to create new todo
Create the function that creates the toDo and optionally pushes it into
a corresponding project. The function is going to be tied to a form that
will allow the user to instanciate a new toDo and select the project it
should belong to. The function resides in a new events.js module where
all the code relating to form submit will be stored.

Create also a new module called checker which tracks the currently
selected project, and the total number of toDos and projects.