Add function to display todos in projects
Add function to display all the todos present in each project with a
click on the project name. As in the allTask function the function
automatically creates a button to add a new to do, in this changes
to the existing project.

Add another fake todo and project (deutschland e rammstein)